pyproject.toml
[project]
name = "buyursin"
version = "0.1.0"
description = ""
authors = [
    {name = "Ibroxim",email = "abdusalomovmailbox@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "aiogram (>=3.0.0)",
    "python-dotenv (>=1.0.0)",
    "sqlalchemy (>=2.0.0)",
    "alembic (==1.16.5)",
    "psycopg2-binary (==2.9.10)",
    "g4f (>=7.0.0,<8.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sentry-sdk (==2.41.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
